requirements.txt
alembic==1.4.3
appdirs==1.4.4
astroid==2.4.2
awscli==1.19.7
Babel==2.9.0
botocore==1.20.7
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
ddtrace==0.47.0
distlib==0.3.1
docutils==0.15.2
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
intervaltree==3.1.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.23
postgres==3.0.0
protobuf==3.15.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pyasn1==0.4.8
pylint==2.5.3
python-dateutil==2.6.0
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
PyYAML==5.3.1
requests==2.25.0
rsa==4.5
s3transfer==0.3.4
six==1.15.0
sortedcontainers==2.3.0
SQLAlchemy==1.4.0
tenacity==7.0.0
toml==0.10.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.11.1
WTForms==2.3.3
python-jose-cryptodome==1.3.2